IS-1631: remove unnecessary files
